+++ business/src/test/java/com/kevinguanchedarias/owgejava/dao/RequirementInformationDaoTest.java
@@ -1,0 +1,54 @@
+package com.kevinguanchedarias.owgejava.dao;
+
+import com.kevinguanchedarias.owgejava.business.FactionBo;
+import com.kevinguanchedarias.owgejava.business.GalaxyBo;
+import com.kevinguanchedarias.owgejava.business.ObjectRelationBo;
+import com.kevinguanchedarias.owgejava.business.SpecialLocationBo;
+import com.kevinguanchedarias.owgejava.business.UpgradeBo;
+import com.kevinguanchedarias.owgejava.repository.ObjectRelationsRepository;
+import com.kevinguanchedarias.owgejava.repository.RequirementInformationRepository;
+import com.kevinguanchedarias.owgejava.repository.RequirementRepository;
+import com.kevinguanchedarias.owgejava.util.ExceptionUtilService;
+import com.kevinguanchedarias.taggablecache.manager.TaggableCacheManager;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+
+import static com.kevinguanchedarias.owgejava.business.RequirementInformationBo.REQUIREMENT_INFORMATION_CACHE_TAG;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+@SpringBootTest(
+        classes = RequirementInformationDao.class,
+        webEnvironment = SpringBootTest.WebEnvironment.NONE
+)
+@MockBean({
+        RequirementInformationRepository.class,
+        SpecialLocationBo.class,
+        FactionBo.class,
+        UpgradeBo.class,
+        GalaxyBo.class,
+        RequirementRepository.class,
+        ObjectRelationBo.class,
+        ObjectRelationsRepository.class,
+        ExceptionUtilService.class,
+        TaggableCacheManager.class
+})
+class RequirementInformationDaoTest {
+    private final RequirementInformationDao requirementInformationDao;
+    private final TaggableCacheManager taggableCacheManager;
+
+    @Autowired
+    RequirementInformationDaoTest(RequirementInformationDao requirementInformationDao, TaggableCacheManager taggableCacheManager) {
+        this.requirementInformationDao = requirementInformationDao;
+        this.taggableCacheManager = taggableCacheManager;
+    }
+
+    @Test
+    void clearCache_should_work() {
+        requirementInformationDao.clearCache();
+
+        verify(taggableCacheManager, times(1)).evictByCacheTag(REQUIREMENT_INFORMATION_CACHE_TAG);
+    }
+}